Add cypress scaffolding (#471)
