+++ .github/workflows/deploy.yml
@@ -119,65 +119,63 @@ jobs:
     outputs:
       application_endpoint: ${{ steps.endpoint.outputs.application_endpoint}}
 
-  # e2e-test:
-  #   name: E2E Integration Tests
-  #   needs: deploy
-  #   if: ${{ github.ref_name != 'production' }}
-  #   runs-on: ubuntu-latest
-  #   steps:
-  #     - uses: actions/checkout@v4
-  #     - name: Run Cypress Tests
-  #       uses: cypress-io/github-action@v4.2.0
-  #       with:
-  #         working-directory: tests/cypress
-  #         spec: |
-  #           tests/e2e/*.spec.js
-  #           tests/e2e/**/*.feature
-  #         browser: chrome
-  #         config: baseUrl=${{ needs.deploy.outputs.application_endpoint }}
-  #         wait-on: ${{ needs.deploy.outputs.application_endpoint }}
-  #         env: true
-  #       env:
-  #         CYPRESS_STATE_USER_EMAIL: ${{ secrets.CYPRESS_STATE_USER_EMAIL }}
-  #         CYPRESS_STATE_USER_PASSWORD: ${{ secrets.CYPRESS_STATE_USER_PASSWORD }}
-  #         CYPRESS_ADMIN_USER_EMAIL: ${{ secrets.CYPRESS_ADMIN_USER_EMAIL }}
-  #         CYPRESS_ADMIN_USER_PASSWORD: ${{ secrets.CYPRESS_ADMIN_USER_PASSWORD }}
-  #     - name: Upload screenshots
-  #       uses: actions/upload-artifact@v3
-  #       if: failure()
-  #       with:
-  #         name: cypress-screenshots
-  #         path: |
-  #           tests/cypress/screenshots/
-  #           tests/cypress/videos/
+  e2e-test:
+    name: E2E Integration Tests
+    needs: deploy
+    if: ${{ github.ref_name != 'production' }}
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Run Cypress Tests
+        uses: cypress-io/github-action@v6
+        with:
+          working-directory: tests
+          spec: |
+            cypress/e2e/*.cy.js
+          browser: chrome
+          config: baseUrl=${{ needs.deploy.outputs.application_endpoint }}
+          wait-on: ${{ needs.deploy.outputs.application_endpoint }}
+          env: true
+        env:
+          CYPRESS_STATE_USER_EMAIL: ${{ secrets.CYPRESS_STATE_USER_EMAIL }}
+          CYPRESS_STATE_USER_PASSWORD: ${{ secrets.CYPRESS_STATE_USER_PASSWORD }}
+          CYPRESS_ADMIN_USER_EMAIL: ${{ secrets.CYPRESS_ADMIN_USER_EMAIL }}
+          CYPRESS_ADMIN_USER_PASSWORD: ${{ secrets.CYPRESS_ADMIN_USER_PASSWORD }}
+      - name: Upload screenshots
+        uses: actions/upload-artifact@v4
+        if: failure()
+        with:
+          name: cypress-screenshots
+          path: |
+            tests/cypress/screenshots/
+            tests/cypress/videos/
 
-  # a11y-tests:
-  #   name: E2E A11y Tests
-  #   needs: deploy
-  #   if: ${{ github.ref_name != 'production' }}
-  #   runs-on: ubuntu-latest
-  #   steps:
-  #     - uses: actions/checkout@v4
-  #     - name: Check Project A11y
-  #       uses: cypress-io/github-action@v4.2.0
-  #       with:
-  #         working-directory: tests/cypress
-  #         spec: tests/accessibility/*.feature
-  #         browser: chrome
-  #         config: baseUrl=${{ needs.deploy.outputs.application_endpoint }}
-  #         wait-on: ${{ needs.deploy.outputs.application_endpoint }}
-  #         env: true
-  #       env:
-  #         CYPRESS_STATE_USER_EMAIL: ${{ secrets.CYPRESS_STATE_USER_EMAIL }}
-  #         CYPRESS_STATE_USER_PASSWORD: ${{ secrets.CYPRESS_STATE_USER_PASSWORD }}
-  #         CYPRESS_ADMIN_USER_EMAIL: ${{ secrets.CYPRESS_ADMIN_USER_EMAIL }}
-  #         CYPRESS_ADMIN_USER_PASSWORD: ${{ secrets.CYPRESS_ADMIN_USER_PASSWORD }}
-  #         RUN_PA11Y: true
-  #     - name: Upload screenshots
-  #       uses: actions/upload-artifact@v3
-  #       if: failure()
-  #       with:
-  #         name: cypress-screenshots
-  #         path: |
-  #           tests/cypress/screenshots/
-  #           tests/cypress/videos/
+  a11y-tests:
+    name: E2E A11y Tests
+    needs: deploy
+    if: ${{ github.ref_name != 'production' }}
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Check Project A11y
+        uses: cypress-io/github-action@v6
+        with:
+          working-directory: tests
+          spec: cypress/e2e/accessibility/*.cy.js
+          browser: chrome
+          config: baseUrl=${{ needs.deploy.outputs.application_endpoint }}
+          wait-on: ${{ needs.deploy.outputs.application_endpoint }}
+          env: true
+        env:
+          CYPRESS_STATE_USER_EMAIL: ${{ secrets.CYPRESS_STATE_USER_EMAIL }}
+          CYPRESS_STATE_USER_PASSWORD: ${{ secrets.CYPRESS_STATE_USER_PASSWORD }}
+          CYPRESS_ADMIN_USER_EMAIL: ${{ secrets.CYPRESS_ADMIN_USER_EMAIL }}
+          CYPRESS_ADMIN_USER_PASSWORD: ${{ secrets.CYPRESS_ADMIN_USER_PASSWORD }}
+      - name: Upload screenshots
+        uses: actions/upload-artifact@v4
+        if: failure()
+        with:
+          name: cypress-screenshots
+          path: |
+            tests/cypress/screenshots/
+            tests/cypress/videos/
