--- tests/cypress/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "cypress-tests",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "cd ../../ && ./dev local && cd -",
-    "test:ci": "cypress install && cypress run --browser chrome --headless",
-    "cypress": "cypress open",
-    "test": "concurrently --kill-others \"yarn start\" \"yarn cypress\""
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@badeball/cypress-cucumber-preprocessor": "^15.1.3",
-    "@cypress-audit/pa11y": "^1.3.1",
-    "@cypress/xpath": "^2.0.3",
-    "@testing-library/cypress": "^9.0.0",
-    "axe-core": "^4.6.3",
-    "concurrently": "^7.6.0",
-    "cypress": "^12.5.1",
-    "cypress-axe": "^1.3.0",
-    "cypress-file-upload": "^5.0.8",
-    "cypress-wait-until": "^1.7.2"
-  },
-  "cypress-cucumber-preprocessor": {
-    "nonGlobalStepDefinitions": true,
-    "stepDefinitions": [
-      "support/step_definitions/**/*.js",
-      "tests/[filepath].stepdef.js",
-      "tests/[filepath]/**/*.stepdef.js"
-    ]
-  },
-  "dependencies": {
-    "cypress-tags": "^1.1.2"
-  }
-}
