+++ tests/cypress/support/commands.js
@@ -1,7 +1,9 @@
-// element selectors
-const cognitoEmailInputField = "//input[@name='email']";
-const cognitoPasswordInputField = "//input[@name='password']";
+// ***** AUTHENTICATION COMMANDS *****
+
+const cognitoEmailInputField = 'input[name="email"]';
+const cognitoPasswordInputField = 'input[name="password"]';
 const cognitoLoginButton = "[data-testid='cognito-login-button']";
+const myAccountButton = '[aria-label="my account"';
 
 const stateUserPassword = Cypress.env("STATE_USER_PASSWORD");
 const adminUserPassword = Cypress.env("ADMIN_USER_PASSWORD");
@@ -30,17 +32,10 @@ const adminUser = {
   password: adminUserPassword,
 };
 
-Cypress.Commands.add("navigateToHomePage", () => {
-  if (cy.location("pathname") !== "/") cy.visit("/");
-});
-
-Cypress.Commands.add("clearSession", () => {
-  cy.session([], () => {});
-});
-
 Cypress.Commands.add("authenticate", (userType, userCredentials) => {
   cy.session([userType, userCredentials], () => {
     cy.visit("/");
+    cy.wait(2000);
     let credentials = {};
 
     if (userType && userCredentials) {
@@ -65,8 +60,8 @@ Cypress.Commands.add("authenticate", (userType, userCredentials) => {
       throw new Error("Must specify either userType or userCredentials.");
     }
 
-    cy.xpath(cognitoEmailInputField).type(credentials.email);
-    cy.xpath(cognitoPasswordInputField).type(credentials.password, {
+    cy.get(cognitoEmailInputField).type(credentials.email);
+    cy.get(cognitoPasswordInputField).type(credentials.password, {
       log: false,
     });
     cy.get(cognitoLoginButton).click();
@@ -76,17 +71,46 @@ Cypress.Commands.add("authenticate", (userType, userCredentials) => {
      * This ensures reused sessions maintain these tokens
      * We expect at least three for the id, access, and refresh tokens
      */
-    cy.waitUntil(() =>
-      cy
-        .window()
-        .then(
-          (window) =>
-            Object.keys(window.localStorage).filter((key) =>
-              key.match(
-                /CognitoIdentityServiceProvider.+(refresh|access|id)Token/
-              )
-            ).length === 3
-        )
-    );
+    cy.wait(4500);
+    cy.get(myAccountButton).should("exist");
   });
 });
+
+// ***** ACCESSIBILITY COMMANDS *****
+
+Cypress.Commands.add("runAccessibilityTests", () => {
+  // run cypress-axe accessibility tests (https://bit.ly/3HnJT9H)
+  cy.injectAxe();
+  cy.checkA11y(
+    null,
+    {
+      values: ["wcag2a", "wcag2aa"],
+      includedImpacts: ["minor", "moderate", "serious", "critical"],
+      rules: {
+        "duplicate-id": { enabled: false },
+      },
+    },
+    terminalLog
+  );
+});
+
+// ***** LOGGING ***** (https://bit.ly/3HnJT9H)
+
+function terminalLog(violations) {
+  cy.task(
+    "log",
+    `${violations.length} accessibility violation${
+      violations.length === 1 ? "" : "s"
+    } ${violations.length === 1 ? "was" : "were"} detected`
+  );
+  // pluck specific keys to keep the table readable
+  const violationData = violations.map(
+    ({ id, impact, description, nodes }) => ({
+      id,
+      impact,
+      description,
+      nodes: nodes.length,
+    })
+  );
+  cy.task("table", violationData);
+}
