+++ tests/cypress/support/e2e.js
@@ -1,0 +1,2 @@
+import "./commands";
+import "cypress-axe";
