+++ tests/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "cypress-tests",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "cypress": "cypress open",
+    "start": "cd ../ && ./dev local && cd -",
+    "test": "concurrently --kill-others \"yarn start\" \"yarn cypress\"",
+    "test:ci": "cypress install && cypress run --browser chrome --headless"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "axe-core": "^4.8.4",
+    "cypress": "^13.6.6",
+    "cypress-axe": "^1.5.0"
+  }
+}
